requirements.txt
streamlit
mediapipe
pillow
numpy
pandas
openpyxl